Set up eslint and pre-commit hooks (#169)
Since all of us agree that product quality should become more prioritized I think that this is a good first step for making out development workflow better.

I decided to use eslint config from @satya164, cause having an experience with this config I need to admit that this config makes my development more confident.

I decided not only to set it up but also force us to follow it. Now it's not possible to commit anything which does not pass eslint checks and tests.

To understand why it's important I added few TODOs, which seem to be legit issues we tended to ignore. Furthermore, we probably won't find them unless setting up this config, so that's the very visible example of positive impact.

Also, I connected the hook with tests. We don't have a lot of them. Maybe we will have more of them someday but even though we were probably not conscious that some of them are broken. Few crashes were easy to fix, but some of them are not super understandable for me and points tt some other issues, which might be more major.

I'm super positive for having it merged ASAP to not come across merge conflicts. Also, it's good moment because we just merged a lot of code and not so many lines of codes are pending to be merged.