Change fast-image patch-package
Android stopped building for me with this patch, with Android Studio pointing to the Nullable import. Using AndroidX import solves this for me, and I think this is how it should be done now (but I am missing context on why it wasn't used here and whether it worked or caused other problems etc.).