Optimism and Arbitrum NFT support (#3784)
* Add simplehash nfts by wallet address handler

* Fetch optimism and arbitrum NFTs

* Setup for Simplehash NFT parser

* Construct permalinks for simplehash nft marketplaces

* Add Quixotic and Stratos to url opener whitelist

* Rename hideOpenSeaAction to hideNftMarketplaceAction

This is to make the name of this parameter more generic
as we now support more than the OpenSea marketplace.

* Lang: lowercase `with` in `View All With Property`

* Rename viewTraitOnOpenseaAction to viewTraitOnNftMarketplaceAction

* Rename viewTraitOnOpensea to viewTraitOnNftMarketplace in attribute Tag

* Retrieve simplehash marketplace name and add as a new field to unique assets

* Setup for using marketplace info in attribute tags

* Hide NFT marketplace action when collection slug is unavailble

Currently, we hide the NFT marketplace action when the NFT is a POAP.

For the Optimism and Arbitrum endpoints, it is possible that
marketplace page information is unavailable. In this case,
we also want to hide the NFT marketplace actions as the pages
will be unavailable.

* Rename handlePressOpensea to handlePressMarketplaceName

* Update lang files for viewing on OpenSea to be more generic

* Add OpenSea as marketplace to OS and Polygon parsers

* Support more generic marketplace collection website

* Show Rainbow Web only for supported networks

We do not support Optimism or Arbitrum NFTs on Rainbow web yet.
This change also includes explicitly passing in the network even
for mainnet NFTs.

* Use NONE constant in NFT brief token info row

* Set floor price to NONE on error or unsupported network

* Show L2 disclaimer for non-mainnet NFTs

* Update Share functionality in unique token expanded state

Since some networks are not supported by Rainbow web yet,
we want to fallback to using the permalink as the Share content.
If a permalink also does not exist, then we do not show the Share
button.

* Show NFT brief info only on supported networks.

We currently do not support last sale and floor price info
for L2s. The entire NFT Brief Info section will be ignored for L2s.

* Add simplehash types

* Support cursor for simplehash handler

* Remove shouldUpdateInBatches flag for unique tokens

With the new additions of unique token fetches, the
shouldUpdateInBatches flag was causing unexpected behavior.

* Fix and simplify action sheet index selection on collection on Android

* Fix and simplify action sheet index selection on asset on Android

* Lint cleanup and review comments

* Linter: update custom ENS unique token creation