Remove non-brief wallet sections data selector (#4498)
* Remove dependence on sections data from non-brief wallet sections

The (non brief) sections data is used to pre-fetch the charts
for the top assets. We can revisit supporting this use case later.
For now, I will remove this as it is not a hard requirement.

* Provide isEmpty from buildBriefWalletSectionsSelector

Move over the isEmpty calculation to the brief selector
and remove dependence of the useWalletSectionsData hook
on buildWalletSectionsSelector

* Remove (non brief) buildWalletSectionsSelector

* Remove unused values from asset selector

Now that the non-brief wallet sections data selector
has been removed, there were some properties it relied on
that are not used anywhere else.