Add shouldListen option to useKeyboardHeight
This change allows conditionally enabling the keyboard height listener.

We need this in order to prevent the keyboardHeight that's applied to the browser search bar from changing as a result of an input within the WebView being focused.

The WebView inserts a toolbar above the keyboard that increases the height of the standard keyboard. Without the use of shouldListen, the browser search bar jumps after being focused if the last focused input was a WebView input.

Also fixes some TS issues in the useKeyboardHook.