Browser upgrades (#6269)
* Dapp browser upgrades

* Remove unneeded whitelisted native text prop

* Clean up SwipeNavigator tab bar components

* Remove BPA wrapper from GestureHandlerButton

Also adds turbo-haptics and an internal scale animation, giving GestureHandlerButton full platform parity

* freezeStore → navigationStore

* Restructure browser gesture handlers, code cleanup

* Make useKeyboardHeight more efficient

* TextShadow: Add support for AnimatedText

* Remove commented code

* Fix lint error

* Remove commented code

* Force a new WebView when the content process dies

Reloading the page by calling reload() on the ref doesn't successfully restart the WebView.

This was causing tabs to occasionally become blank after the app sat backgrounded for long enough, and upon returning to the app they'd be stuck.

This change forces a new WebView by providing a new key, which should resolve the issue.

* Minor gesture cleanup

* Add common gesture options to GestureHandlerButton

* Bump packages

* Fix manual activation bug in RNGH

Fixes a bug that exists with manually activated gestures where when the app returns to the foreground, the first interaction with the manually activated handler doesn't trigger any events

* Fix type issue in react-native-view-shot

RN was complaining that this iOS code was not Android compliant

* Only render WebViewBorder on iOS, misc. cleanup

* Adjust ScrollView/tabs hierarchy, props cleanup

* Bump Skia

* Remove activeTabInfo.value.tabId

No longer needed as there's a dedicated activeTabId Shared Value

* Make tab switching lighter, add getTabInfo helper

A few tab switching styles were applied for all tabs when tab switching, when only the active tab and the tabs to the left and right need full-size tab styles

* Smooth out height animations

* Clean up scripts, clamp tab height

* Clean up WebView handlers

* Improve tab animation performance

- Removed unnecessary isSwitchingTabs (`tabViewGestureState` serves the same purpose)

- Removed a wrapper view around each tab and merged most animated tab styles into `animatedWebViewStyle`

- In `animatedWebViewStyle`, now passing values to the style generation worklets instead of passing the Shared Values themselves

* Make screenshot flicker less likely, improve tab switching reset

- Now setting the JS-side active tab index first so it gets a tiny head start (the JS and SV active indexes inform which tab activates and when tab screenshots are shown/hidden)

- Now resetting tab switching state *before* updating `activeTabId` and `animatedActiveTabIndex` (instead of after) — this fixes an edge case where the switch animation would occasionally be skipped, if you switched tabs again at the exact moment a previous tab switch ended

* Make isNewIndexValid a boolean

* Fix bad worklet pattern

Because getTabInfo is stable and the SVs used within getTabInfo weren't visible to the useAnimatedStyle hooks, styles weren't recomputing for changes in getTabInfo's input values — same idea applies to useDerivedValue, useAnimatedReaction prepare functions, etc.