Stabilize transaction watchers
In certain cases when swapping or creating transactions, a runaway retry loop would occur driven by a combination of transaction fetch failures and the useEffect logic in `useTransactionWatcher`. Seems to have been a mix of instability in `processTransactions` and the abort logic.

This simplifies the structure of the retry tick and generates a key from the pending or mined transactions, which is used to trigger the useEffect.

It uses refs for the watch function and the transactions data to further limit instability and avoid stale closures.