Adjust condition to work around Legend List bug
Seems to be a bug in Legend List where in some cases the list doesn't update in response to `extraData` changing, in this case when navigating from the main perps screen → tapping a token in the Markets section → tapping Open Position.

So this effectively defaults the list to be in the new position state rather than in the search state, so that when it fails to update it's already in the correct state. Seems to be the only affected case.