Perps polish, state cleanup (#6866)
* Add virtual navigator system

* Allow extra method bundling in `createStoreActions`

* Optimize derived query store params

Ensures that derived stores used as query store params are subscribed to **before** their `getState` method is called, so that the result of that initial call is cached. This avoids a second unnecessary invocation of the derive function when derived query params are initialized.

* Polish wallet screen perps UI

* Add perps navigator

* Add `THICKER_BORDER_WIDTH`

* Add Hyperliquid charts prefetching, fix chart type handling

* State cleanup, UI polish

* Fix price change colors, active route tracking

Need to use the perps navigation store directly here for now

* Optimize active position check

* Fix disabled perps market row spacing

* Fix total account balance

* Fix division by zero

* Fix deposit/withdrawal screen slider crash

* Quick fix for close position sheet value updates

Will improve further with the other slider fixes — should just be derived from the slider position

* Fix stuck slider after press and hold, remove unneeded 'worklet' tags

* Ensure 2 decimal places when there are decimal places

e.g. $2.4 → $2.40

* Ensure consistent decimal precision

* Restore 2 decimal place minimum

* Remove unnecessary external call for nav state

* Bump Legend List

* Adjust condition to work around Legend List bug

Seems to be a bug in Legend List where in some cases the list doesn't update in response to `extraData` changing, in this case when navigating from the main perps screen → tapping a token in the Markets section → tapping Open Position.

So this effectively defaults the list to be in the new position state rather than in the search state, so that when it fails to update it's already in the correct state. Seems to be the only affected case.

* Improve `SmoothPager` allowed bounds logic

* Adjust disabled market row spacing

* Only update params if specified