TL updates with transactions fixes (#6846)
* WIP

* Remove redundant code

* Add support for tlv2 envs

* TEMP: override for tenderly on base

* Allow creating with no token logo

* Update package.json to use private tl-v2 npm package and alias existing usage of tl in code

* Add resolution for tanstack query-core

* Revert "TEMP: override for tenderly on base"

This reverts commit 732c7452b33b61118dd0bf0ad8b139725cde470d.

* Update yarn.lock

* Add fallback KOTH coin icons

* Fix fallback icons

* Re-add price to token launcher footer

* Temp: re-add dev section env vars analyzer

* Adjust KOTH explain sheet copy

* Add token search url to analyze env vars

* Add fallback if king token has no coin icon

* Adjust leaderboard item badge style

* Update Legend List, minor list optimizations

* Clean up pending transaction and nonce stores

Also migrates both to `createRainbowStore`

* Make `clearMinedTransactions` idempotent

* Patch abort leaks in the addys client singleton

This was causing cross-contaminated aborts in unrelated addys requests

* Reduce type assertions, remove false async tag

* Stabilize transaction watchers

In certain cases when swapping or creating transactions, a runaway retry loop would occur driven by a combination of transaction fetch failures and the useEffect logic in `useTransactionWatcher`. Seems to have been a mix of instability in `processTransactions` and the abort logic.

This simplifies the structure of the retry tick and generates a key from the pending or mined transactions, which is used to trigger the useEffect.

It uses refs for the watch function and the transactions data to further limit instability and avoid stale closures.

* Normalize transaction statuses, bypass React Query for transaction fetches

The addys transaction endpoint seems to be inconsistent in where it places the transaction status, so this adds normalization logic to account for it.

This also bypasses the React Query client when fetching transactions, as it's unnecessary, and uses clearer naming for the transactions fetch functions.

* Remove unnecessary useMemo

* Increase `NODE_ACK_DELAY` to 2500ms

Unideal but temporarily patches failed swaps that currently occur on Base in the approve+swap rap flow

* [Revert] Add temporary logging

* Memoize watcher transactions key

* Use higher node acknowledgement delay for Base

* Require token logo outside of TestFlight and dev mode

* Remove transactions logging code

* Fix speed up and cancel

* Apply the correct asset when speeding up or canceling, hide the value field in the speed up and cancel sheets as it's not correctly formatted nor useful

* Remove env vars dev menu

* Fix lint error in old KOTH card

* Fix missing `ethers` import

* Improve `fetchedTransaction` merge logic to account for `type` and `title`

* Fix LegendList padding styles, enable recycling in KOTH list

* Always prefer the local transaction `type`

* Simplify `fetchedTransaction` merging

* Optimize `pendingTransactions` subscriptions

Restricts re-renders to when the pending transactions data meaningfully changes by ignoring outer object reference changes. Should be further optimized at the point where these no-op updates/reference changes are occurring.

* Move pending transactions optimizations to the setter

* Remove ability to pass an `abortController` to `getAddysHttpClient`

* Pass the abort signal through to `fetchRawTransaction`

* Hook up the positions abort signal

* Remove transactions key logic from `useTransactionWatcher`

No longer needed as the instability in the transactions data was resolved at the source. Also now depending on the `watchFunction` directly.

* Distinguish pending swaps from bridges

Previously bridges were classified as swaps, which caused the transaction row to show a "Swapping" label for bridges while pending, rather than a "Bridging" label.

* Prefer the fetched transaction upon confirmation (in most cases)

Special cases a few transaction types for which the fetched transaction contains bad data — unideal but works for now.

* bump iOS and Android to v2.0.2

* Remove unused TL constants, adjust creator bps

* Ensure correct pending transaction ordering

* Always show rank coin icon border

* Remove higher node ack delay on Base

* Fix comment typo

* Version bump tlv2

* Hide discover cards

* Remove unused scope in yarnrc

* Bump iOS and Andoid to v.2.0.5

* Update heap space

* Update heap space for ios e2e runs

---------

Co-authored-by: Christian Baroni <7061887+christianbaroni@users.noreply.github.com>
Co-authored-by: Ibrahim Taveras <ibrahim@rainbow.me>